requirements_sphinx.txt
izi-core==1.5.3
djangorestframework>=3.7.0
django-nose
django-haystack>=2.7.0.dev
Django>=1.11,<2.0
